Split iterator sources into different modules
